% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source.R
\name{create_source}
\alias{create_source}
\title{Define a source for a Socrata dataset revision}
\usage{
create_source(
  revision_response,
  username,
  password,
  source_type = "upload",
  source_parse = TRUE
)
}
\arguments{
\item{revision_response}{The named list returned by \code{open_revision}.}

\item{username}{A Socrata username or API Key ID}

\item{password}{A Socrata password or API Key Secret}

\item{source_type}{a character string indicating the type of source: 'upload' for files uploaded
to Socrata is the only supported option at this time.}

\item{source_parse}{logical, indicating if the source is a parseable filetype.

Filetypes parseable by Socrata include Comma-separated values (CSV), Tab-separated values (TSV),
Microsoft Excel (XLS), Microsoft Excel (OpenXML), ZIP archive (shapefile), JSON format (GeoJSON),
GeoJSON format, Keyhole Markup Language (KML), Zipped Keyhole Markup Language (KMZ)}
}
\value{
A named list with the asset id, the revision sequence id,
the url of the revision, the status code returned while
creating the source, and the full response body.
}
\description{
Defines a source for a Socrata dataset revision, including that the new data
will be of \code{source_type} "upload" and whether the new data is parseable or not by Socrata
with \code{source_parse}. Defining a source is a prerequite to actually sending new data to the
revision in \code{upload_to_source}.
}
