% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_revision.R
\name{delete_revision}
\alias{delete_revision}
\title{Delete a Socrata revision}
\usage{
delete_revision(domain, dataset_id, revision_id, email, password)
}
\arguments{
\item{domain}{A Socrata domain name.}

\item{dataset_id}{A Socrata dataset identifier, or four-by-four.}

\item{revision_id}{An integer identifying the revision you wish to delete.}

\item{email}{A Socrata username or API Key ID}

\item{password}{A Socrata password or API Secret ID}
}
\value{
A HTTP status code for the delete request.
}
\description{
Deletes a Socrata dataset revision, identified by \code{dataset_id} and
\code{revision_id}, the revision sequence number. Revisions are numbered
sequentially from 0, the very first revision, and increment by 1 for each
new revision.
}
\details{
A revision id can be identified by looking at the end of the URL of a draft,
i.e. 4 in "/dataset/dataset-name/abcd-1234/revisions/4", or by making a GET
request to "https://yourdomain.gov/api/publishing/v1/revision/abcd-1234",
with your dataset id, finding the most recent or relevant revision, and
locating the \code{revision_seq} number.
}
