% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_domain.R
\name{validate_domain}
\alias{validate_domain}
\title{Validates the format of a provided Socrata domain}
\usage{
validate_domain(domain)
}
\arguments{
\item{domain}{A Socrata hostname.}
}
\value{
A Socrata hostname.
}
\description{
Validates the format of a provided Socrata domain
}
