% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_revision.R
\name{open_revision}
\alias{open_revision}
\title{Open a revision on a Socrata dataset}
\usage{
open_revision(domain, dataset_id, action_type, email, password)
}
\arguments{
\item{domain}{The domain name of the Socrata site}

\item{dataset_id}{The dataset id, or four-by-four, of the Socrata dataset}

\item{action_type}{The type of revision you want to create: "update",
"replace", or "delete"}

\item{email}{Your Socrata username or API Key ID}

\item{password}{Your Socrata password or API Key Secret}
}
\value{
A named list of the dataset id, the url of the new revision,
and the status code returned
}
\description{
Opens a revision (draft) on a Socrata dataset. When opening a
revision on an existing dataset, provide the dataset id as well as an action
type. The action type determines the type of revision: an "update", "replace",
or "delete" revision. An "update" revision will utilize a row identifier, if
one is set on the Socrata dataset, to update existing rows and append new
rows. A "replace" will perform a complete replace of the existing dataset
with the file you are sending with the revision A "delete" action type
requires a row identifier to be have been set on the Socrata dataset, as
the revision will delete rows in the published dataset that match to rows
provided in the file in the revision.

This function simply opens a revision and sets the action type of that
revision.
}
